services/claude_llm_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-interviewer-claude-llm-agent"
version = "1.0.0"
description = "Claude LLM processing MCP Mesh service for AI Interviewer system"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AI Interviewer Team", email = "team@aiinterviewer.com"}
]
keywords = [
    "mcp", "mesh", "llm", "claude", "anthropic", "ai-processing", 
    "ai-interviewer", "text-analysis"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic"
]
requires-python = ">=3.8"
dependencies = [
    # MCP Mesh framework - using published version
    "mcp-mesh>=0.5.0,<0.6.0",
    "fastmcp>=0.9.0",
    
    # Claude API integration
    "anthropic>=0.8.0",
    
    # Data validation and processing
    "pydantic>=2.0.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
    "tenacity>=8.0.0",  # Retry logic for API calls
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    
    # Code quality
    "black>=23.0.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=6.0.0,<7.0.0",
    "mypy>=1.0.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/ai-interviewer/ai-interviewer"
Repository = "https://github.com/ai-interviewer/ai-interviewer"
Issues = "https://github.com/ai-interviewer/ai-interviewer/issues"

[project.scripts]
claude-llm-agent = "claude_llm_agent.__main__:main"

[tool.setuptools.packages.find]
include = ["claude_llm_agent*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
claude_llm_agent = ["py.typed", "*.json", "*.yaml", "*.yml"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

# isort configuration  
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

# MyPy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "anthropic.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests", 
    "unit: marks tests as unit tests",
]